nlp-service/requirements.txt
fastapi
uvicorn
transformers
torch
numpy
python-dotenv
pymupdf
pillow
pytesseract
requests
python-multipart
protobuf
sentencepiece